+++ zipapps/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from .main import create_app, __version__, ZipApp
+from .main import create_app, __version__, ZipApp, pip_install_target
 from .activate_zipapps import activate
 
-__all__ = ["create_app", "activate", "__version__", "ZipApp"]
+__all__ = ["create_app", "activate", "__version__", "ZipApp", "pip_install_target"]